+++ include/sdcard.hpp
@@ -7,21 +7,16 @@ class sdcard
 {
 private:
     const char *file = "macros.txt";
-    enum sdstatus
-    {
-        notconnected,
-        connected,
-        no_file
-    };
     char *internal_str;
     int times_called;
     void refereshFileInfo();
+    bool sdconnected;
 
 public:
+    bool init();
+    bool checkConnection();
     int max_lines, max_str_len;
     char *readLine(int line);
-    int sdconnected;
-    bool init();
 };
 
 extern sdcard sd;